refactor(core): move api-description to api-elements
