+++ packages/api-elements/lib/elements/copy.js
