+++ packages/api-elements/lib/elements/resource.js
