+++ packages/api-elements/lib/elements/transition.js
