--- packages/minim-api-description/CHANGELOG.md
@@ -1,104 +1,0 @@
-# Minim API Description
-
-## 0.9.1 (2019-03-26)
-
-### Enhancements
-
-- `Element.prototype.valueOf` can now accept an array of elements to be used
-  for handling references.
-
-## 0.9.0 (2019-02-26)
-
-### Breaking
-
-- Node 4 is no longer supported.
-- The package no longer contains a browser bundle of minim-api-description.
-
-### Enhancements
-
-- Update minim to 0.22.1.
-- Introduced .valueOf method reducing an API Element from its type to a normalised value.
-
-## 0.8.1
-
-### Enhancements
-
-- Improved documentation coverage of JSDoc comments for methods.
-
-## 0.8.0
-
-- Added `EnumElement`
-
-## 0.7.0
-
-- Added browser build distribution
-
-## 0.6.0
-
-- Updated to support Minim 0.19.
-
-## 0.5.0
-
-- Update to follow Refract 1.0 / API Elements 1.0.
-- Compatibility with Minim 0.18.
-
-## 0.4.1 (2017-06-29)
-
-- We don't try to re-define `copy` on an Element when `copy` already exists.
-  Such as when you create multiple namespaces of minim-api-description.
-
-## 0.4.0 (2017-06-29)
-
-- Updated to support minim 0.17.1.
-
-## 0.3.1 (2017-05-12)
-
-- Added API metadata convenience function
-
-## 0.3.0 (2017-04-04)
-
-- Make sure all convenience functions return an element
-
-## 0.2.0 (2017-03-30)
-
-- Updated babel-runtime dependency to v6
-- Dropped support for node 0.12 and 0.10
-
-## 0.1.8 (2016-09-02)
-
-### Bug Fixes
-
-- A data structure element may contain an array as it's content.
-
-## 0.1.7 (2016-06-30)
-
-- Added an `extension` element
-
-## 0.1.6 (2016-05-24)
-
-- Added a `copy` getter to all array elements
-
-## 0.1.5 (2016-05-04)
-
-- Added support for authScheme related elements and attributes
-
-## 0.1.4 (2016-04-28)
-
-- Upgrade minim peerDependency version to 0.14.0
-
-## 0.1.3 (2016-01-14)
-
-- Move minim to devDependency and update it to 0.13.0
-
-## 0.1.2 (2015-09-24)
-
-- Documentation updates
-- Fix handling of Category attribute `meta`, which is a refracted element.
-
-## 0.1.1 (2015-09-07)
-
-- Update build system to use [Peasant](https://github.com/danielgtaylor/peasant)
-
-## 0.1.0 (2015-09-04)
-
-- Initial release.
