--- packages/minim-api-description/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "minim-api-description",
-  "version": "0.9.1",
-  "description": "Minim API Description Namespace",
-  "author": "Apiary.io <support@apiary.io>",
-  "license": "MIT",
-  "main": "./lib/api-description.js",
-  "homepage": "https://github.com/apiaryio/api-elements.js/tree/master/packages/minim-api-description",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/apiaryio/api-elements.js.git",
-    "directory": "packages/minim-api-description"
-  },
-  "scripts": {
-    "lint": "eslint .",
-    "lint:fix": "eslint . --fix",
-    "test": "mocha"
-  },
-  "devDependencies": {
-    "chai": "^4.1.2",
-    "eslint": "^5.9.0",
-    "minim": "^0.23.1",
-    "mocha": "^5.0.2"
-  },
-  "engines": {
-    "node": ">=6"
-  }
-}
